Swing/Pull Velocity and Speed
Developer can now choose between Swing/Pull Velocity and Speed. When Velocity is 0.0f, the velocity of the character will be set to the speed. If the Velocity value is set, this amount will be added to the rigidbodies speed as velocity. The speed value is now the maximum speed the character should reach with the regarding velocity values.