Player now set in correct State, after reaching object with Pulling
Before the player was always set in lastActionState, which obviously was Hooked, while Pulling to the object. But since the Rope breaks after reaching the object, he should be set to Free after the Falling check. Thus the lastActionState is now manually set to Free before doing the Falling Check. Setting the ActionState to Free with the SetActionState-Method, the lastSavePosition would be set falsely above a pit, if the players Pulling ends there