Base Statemachine
The base of the state machine is standing. New States have been added, some states have been changed.
No more handling between different states, which state should be checked, since that should be dependent on the multiple entities. The Handling for that now runs in the explicit entity itself (i.e. EntityPlayer handles when the check for movement, dash etc. will be done).